Added error checking for winsock and scanf. Changed fseek to use the 64bit version where available. Added 64bit compile flag for vs2010 projects. Moved socket status variable from Protocol Base to the Socket class

+++ include/SmartPointer.h
@@ -114,7 +114,7 @@ public:
 	}
 	
 	// Smart pointer array accessor
-	T& operator[] ( int x ) {
+	T& operator[] ( intptr_t x ) {
 		return pointer[x];
 	}
 private:
